+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_functions.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/docker/go-connections/nat"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_log_streaming_readcloser"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_operation_parallelizer"
@@ -14,7 +15,7 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/exec_result"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion_volume"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/stacktrace"
@@ -80,6 +81,7 @@ var commandToRunWhenCreatingUserServiceShell = []string{
 	"-c",
 	"if command -v 'bash' > /dev/null; then echo \"Found bash on container; creating bash shell...\"; bash; else echo \"No bash found on container; dropping down to sh shell...\"; sh; fi",
 }
+
 
 type userServiceDockerResources struct {
 	// This will never be nil because a user services is represented ONLY by a container in Docker
@@ -147,7 +149,7 @@ func (backend *DockerKurtosisBackend) RegisterUserService(ctx context.Context, e
 	return registration, nil
 }
 
-func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, enclaveId enclave.EnclaveID, serviceGuid service.ServiceGUID, containerImageName string, privatePorts map[string]*port_spec.PortSpec, entrypointArgs []string, cmdArgs []string, envVars map[string]string, filesArtifactVolumeMountDirpaths map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]string, ) (*service.Service, error, ) {
+func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, enclaveId enclave.EnclaveID, serviceGuid service.ServiceGUID, containerImageName string, privatePorts map[string]*port_spec.PortSpec, entrypointArgs []string, cmdArgs []string, envVars map[string]string, filesArtifactVolumeMountDirpaths map[files_artifact_expansion.FilesArtifactExpansionGUID]string, ) (*service.Service, error, ) {
 	backend.serviceRegistrationMutex.Lock()
 	defer backend.serviceRegistrationMutex.Unlock()
 
@@ -240,10 +242,34 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 	if cmdArgs != nil {
 		createAndStartArgsBuilder.WithCmdArgs(cmdArgs)
 	}
+
 	if filesArtifactVolumeMountDirpaths != nil {
+
+		// CREATE FILTER SET FOR FILES ARTIFACT EXPANSION GUIDS AND SERVICE GUID
+		filterFilesArtifactExpansionGuids := map[files_artifact_expansion.FilesArtifactExpansionGUID]bool{}
+		for guid, _ := range filesArtifactVolumeMountDirpaths {
+			filterFilesArtifactExpansionGuids[guid] = true
+		}
+		filters := files_artifact_expansion.FilesArtifactExpansionFilters{
+			GUIDs: filterFilesArtifactExpansionGuids,
+			ServiceGUIDs: map[service.ServiceGUID]bool{
+				serviceGuid: true,
+			},
+		}
+		// GET MATCHING EXPANSION OBJECTS AND RESOURCES FOR FILTERS
+		filteredObjectsAndResources, err := backend.getMatchingFilesArtifactExpansionObjectsAndDockerResources(ctx, enclaveId, &filters)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Failed to get expansion volumes for filters '%+v'", filters)
+		}
+		// ITERATE THROUGH MATCHING OBJECTS AND PLACE VOLUME NAMES FROM DOCKER OBJECTS INTO MAP FOR VOLUME TO PATH MAPPING
 		filesArtifactVolumeMountDirpathStrs := map[string]string{}
-		for filesArtifactVolumeName, mountDirpath := range filesArtifactVolumeMountDirpaths {
-			filesArtifactVolumeMountDirpathStrs[string(filesArtifactVolumeName)] = mountDirpath
+		for filesArtifactGUID, objectsAndResources := range filteredObjectsAndResources {
+			resources := objectsAndResources.dockerResources
+			if resources == nil {
+				return nil, stacktrace.Propagate(err, "Found expansion with no docker resources for files artifact expansion '%v'", filesArtifactGUID)
+			}
+			volume := resources.volume
+			filesArtifactVolumeMountDirpathStrs[volume.Name] = filesArtifactVolumeMountDirpaths[filesArtifactGUID]
 		}
 		createAndStartArgsBuilder.WithVolumeMounts(filesArtifactVolumeMountDirpathStrs)
 	}
@@ -330,6 +356,7 @@ func (backend *DockerKurtosisBackend) GetUserServiceLogs(
 	//TODO use concurrency to improve perf
 	successfulUserServicesLogs := map[service.ServiceGUID]io.ReadCloser{}
 	erroredUserServices := map[service.ServiceGUID]error{}
+	shouldCloseLogStreams := true
 	for guid, resourcesForService := range allDockerResources {
 		container := resourcesForService.container
 		if container == nil {
@@ -337,15 +364,29 @@ func (backend *DockerKurtosisBackend) GetUserServiceLogs(
 			continue
 		}
 
-		readCloserLogs, err := backend.dockerManager.GetContainerLogs(ctx, container.GetId(), shouldFollowLogs)
+		rawDockerLogStream, err := backend.dockerManager.GetContainerLogs(ctx, container.GetId(), shouldFollowLogs)
 		if err != nil {
 			serviceError := stacktrace.Propagate(err, "An error occurred getting logs for container '%v' for user service with GUID '%v'", container.GetName(), guid)
 			erroredUserServices[guid] = serviceError
 			continue
 		}
-		successfulUserServicesLogs[guid] = readCloserLogs
+		defer func() {
+			if shouldCloseLogStreams {
+				rawDockerLogStream.Close()
+			}
+		}()
+
+		demultiplexedLogStream := docker_log_streaming_readcloser.NewDockerLogStreamingReadCloser(rawDockerLogStream)
+		defer func() {
+			if shouldCloseLogStreams {
+				demultiplexedLogStream.Close()
+			}
+		}()
+
+		successfulUserServicesLogs[guid] = demultiplexedLogStream
 	}
 
+	shouldCloseLogStreams = false
 	return successfulUserServicesLogs, erroredUserServices, nil
 }