+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_manager/consts/kubernetes_consts.go
@@ -14,6 +14,7 @@ const (
 	RolesKubernetesResource                  = "roles"
 	RoleBindingsKubernetesResource           = "rolebindings"
 	PodsKubernetesResource                   = "pods"
+	PodExecsKubernetesResource                   = "pods/exec"
 	ServicesKubernetesResource               = "services"
 	PersistentVolumeClaimsKubernetesResource = "persistentvolumeclaims"