+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_annotation_key_consts/kubernetes_annotation_key_consts_test.go
@@ -1,4 +1,4 @@
-package annotation_key_consts
+package kubernetes_annotation_key_consts
 
 import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/kubernetes_annotation_key"
@@ -12,7 +12,7 @@ var labelKeyStrsToEnsure = map[string]string{
 }
 
 var labelKeysToEnsure = map[*kubernetes_annotation_key.KubernetesAnnotationKey]string{
-	PortSpecsAnnotationKey: "kurtosistech.com/ports",
+	PortSpecsKubernetesAnnotationKey: "kurtosistech.com/ports",
 }
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! IMPORTANT !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!