+++ container-engine-lib/docs/changelog.md
@@ -6,11 +6,39 @@
 * Renamed `KurtosisBackend.CopyFromUserService` -> `CopyFilesFromUserService`
     * Users should update their usages accordingly
 
+# 0.28.0
+### Fixes
+* Fixed bug related to having two annotations-key-consts for Kubernetes objects
+
+### Breaking Changes
+* `Module.GetPublicIP` is renamed to `GetMaybePublicIP`
+    * Remediation: switch to new version
+* `Module.GetPublicPorts` renamed to `GetMaybePublicPorts`
+    * Remediation: switch to new version
+* `Module.GetPublicIp` renamed to `Module.GetPublicIP`
+    * Remediation: switch to new version
+    
+# 0.27.0
+### Breaking Changes
+* Unified file expansion volume and expanders into one interface with two associated methods (instead of two interfaces and four methods)
+
+### Changes
+* Switched the API container to get its port info from the serialized port specs on the Kubernetes service
+
+### Fixes
+* Fixed the engine container being listed as running if the engine service had selectors defined
+* Switched our UUID generation to be fully random (v4) UUIDs rather than v1
+* Fixed our exec command
+* Fixed API containers and engines not being able to run pod exec commands
+* Fixed a bug that caused service registration to fail
+* Fixed the user services and modules Docker log streams not actually coming back demultiplexed
+
 # 0.26.1
+### Features
+* Added `KubernetesKurtosisBackend.GetModuleLogs`
 
 # 0.26.0
 ### Features
-* Added `KubernetesKurtosisBackend.GetModuleLogs`
 * Added the functionality to wait until the GRPC port is available before returning when creating `Engines`, `API containers` and `Modules`  
 * Added `KubernetesKurtosisBackend.CreateModule`, `KubernetesKurtosisBackend.GetModules`, `KubernetesKurtosisBackend.StopModules` and `KubernetesKurtosisBackend.DestroyModules`
 * Added `ForModulePod` and `ForModuleService` to `KubernetesEnclaveObjectAttributesProvider`